DROTH-3937 Add check for empty ID list in deletion

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/dao/RoadLinkReplacementWorkListDAO.scala
@@ -45,7 +45,9 @@ class RoadLinkReplacementWorkListDAO {
   }
 
   def deleteFromWorkList(idsToDelete: Set[Long]) = {
-    sqlu"""DELETE FROM road_link_replacement_work_list WHERE id IN (#${idsToDelete.mkString(",")})""".execute
+    if(idsToDelete.nonEmpty) {
+      sqlu"""DELETE FROM road_link_replacement_work_list WHERE id IN (#${idsToDelete.mkString(",")})""".execute
+    }
   }
 
 }
